requirements.txt
alembic==1.6.3
argcomplete==1.12.0
attrs==20.3.0
autopep8==1.5.5
bcrypt==3.1.7
Beaker==1.10.0
beautifulsoup4==4.9.3
bidict==0.21.2
blivet==3.3.3
blivet-gui==2.2.1
Bottleneck==1.2.1
Brlapi==0.8.2
cached-property==1.5.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
chrome-gnome-shell==0.0.0
cli-helpers==2.1.0
click==7.1.2
configobj==5.0.6
cryptography==3.4.6
cupshelpers==1.0
cycler==0.10.0
dasbus==1.4
dbus-python==1.2.16
decorator==4.4.2
distro==1.5.0
dnspython==1.16.0
docker==4.4.4
docker-compose==1.28.6
docker-pycreds==0.4.0
dockerpty==0.4.1
docopt==0.6.2
eventlet==0.31.0
Flask==1.1.2
Flask-Cors==3.0.10
Flask-Migrate==3.0.0
Flask-SocketIO==5.0.1
Flask-SQLAlchemy==2.5.1
fros==1.1
gevent==20.9.0
gpg==1.15.1
greenlet==1.1.0
gunicorn==20.0.4
humanize==0.5.1
idna==2.10
importlib-resources==1.0.2
iredis==1.9.1
itsdangerous==1.1.0
Jinja2==2.11.3
jsonschema==3.2.0
kiwisolver==1.3.1
langtable==0.0.54
libcomps==0.1.15
libvirt-python==7.0.0
lit==12.0.0rc1
lxml==4.6.3
Mako==1.1.4.dev0
Markdown==3.3.3
MarkupSafe==1.1.1
matplotlib==3.4.2
mdv==1.7.4
meson==0.56.2
mistune==0.8.4
mycli==1.24.1
nftables==0.1
numexpr==2.7.1
numpy==1.20.1
olefile==0.46
pandas==1.2.1
pandas-datareader==0.9.0
paramiko==2.7.2
Paste==3.5.0
pendulum==2.1.2
pexpect==4.8.0
pgcli==1.9.0
pgspecial==1.11.5
pid==2.2.3
Pillow==8.1.2
ply==3.11
productmd==1.33
prompt-toolkit==3.0.18
psycopg2-binary==2.8.6
ptyprocess==0.6.0
pwquality==1.4.4
pyaes==1.6.1
pyasn1==0.4.8
pycairo==1.20.0
pycodestyle==2.6.0
pycparser==2.20
pycrypto==2.6.1
pycups==2.0.1
pycurl==7.43.0.6
pyenchant==3.2.0
Pygments==2.9.0
PyGObject==3.40.1
pykickstart==3.32
PyMySQL==1.0.2
PyNaCl==1.4.0
pyOpenSSL==20.0.1
pypandoc==1.5
pyparsing==2.4.7
pyparted==3.11.7
pyperclip==1.8.2
PyQt5==5.15.3
PyQt5-Qt==5.15.2
PyQt5-sip==12.8.1
pyrsistent==0.17.3
PySocks==1.7.1
python-augeas==0.5.0
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
python-engineio==4.0.0
python-Levenshtein==0.12.0
python-meh==0.50
python-socketio==5.0.4
pytz==2021.1
pytzdata==2020.1
pyudev==0.22.0
pyxdg==0.27
PyYAML==5.4.1
redis==3.5.3
regex==2021.4.4
requests==2.25.1
requests-file==1.5.1
requests-ftp==0.3.1
rpm==4.16.1.3
scipy==1.6.2
selinux==3.2
sepolicy==3.2
setools==4.4.0
simpleline==1.8
six==1.15.0
slip==0.6.4
slip.dbus==0.6.4
sos==4.1
soupsieve==2.2
SQLAlchemy==1.4.15
sqlparse==0.3.1
swtpm-localca==0.5.2
swtpm-setup==0.5.2
systemd-python==234
tables==3.6.1
tabulate==0.8.9
Tempita==0.5.1
terminaltables==3.1.0
texttable==1.6.2
toml==0.10.2
tqdm==4.56.0
urllib3==1.25.10
wcwidth==0.1.9
websocket-client==0.57.0
Werkzeug==1.0.1
xlrd==2.0.1
xlwt==1.1.2
zope.event==4.5.0
zope.interface==5.1.2
